feat: 사용자 닉네임 랭크 조회 API

+++ backend/gumison/src/test/java/com/ssafy/gumison/common/util/RankProviderTest.java
@@ -25,7 +25,7 @@ class RankProviderTest {
     //given
     final int SIZE = userRepositorySupport.findNicknamesAndExpAll().size();
     //when
-    int returnSize = rankProvider.loadAllUserExpIntoRankZSet();
+    Long returnSize = rankProvider.loadAllUserExpIntoRankZSet();
     //then
     assertEquals(SIZE, returnSize);
   }
